\name{setsim-package}
\alias{setsim-package}
\alias{setsim}
\docType{package}
\title{Software tool for the visualization of confidence regions in statistical models}

\description{
A collection of R functions for visualizing the targeted likelihood confidence regions (intervals) for any function of the parameters in a variety of regular parametric models.
}

\details{
\tabular{ll}{
Package: \tab setsim\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2012-02-15\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}}
\author{
Boxuan Cui <cui@math.umass.edu>, Philip Pham <philip.pham2@gmail.com>, Daeyoung Kim <daeyoung@math.umass.edu>

Maintainer: Boxuan Cui <cui@math.umass.edu>, Philip Pham <philip.pham2@gmail.com>
}

\references{
Daeyoung Kim and Bruce G. Lindsay (2011). Using confidence distribution sampling to visualize confidence sets. \emph{Statistica Sinica} 21(2), 923--948.
}